chore: Only allow one UV setup to proceed at a time.
